Move is_char specializations to xchar.h
